move TelegramContext up
